Fixed some imports due to moving api models
