+++ python/tests/plugins/test_aac_rest_api_impl.py
@@ -5,7 +5,7 @@ import json
 
 from aac.lang.active_context_lifecycle_manager import get_active_context
 from aac.plugins.rest_api.aac_rest_app import app
-from aac.plugins.rest_api.definition_model import to_definition_model
+from aac.plugins.rest_api.models.definition_model import to_definition_model
 from aac.spec import get_aac_spec
 
 from tests.active_context_test_case import ActiveContextTestCase
